fix: optimize existing reliability issues
