+++ internal/engine/engine.go
@@ -3,10 +3,8 @@ package engine
 import (
 	"fmt"
 	"slices"
-	"sort"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/yuluo-yx/typo/internal/commands"
 	"github.com/yuluo-yx/typo/internal/parser"
@@ -48,26 +46,7 @@ type distanceMatchConfig struct {
 	similarityThreshold float64
 }
 
-type commandRuneCandidate struct {
-	name  string
-	runes []rune
-}
-
-type commandCandidateIndex struct {
-	all       []commandRuneCandidate
-	byRuneLen map[int][]commandRuneCandidate
-	minLen    int
-	maxLen    int
-}
-
-const (
-	fixSourceParser               = "parser"
-	fixSourceHistory              = "history"
-	fixSourceDistance             = "distance"
-	longOptionSimilarityThreshold = 0.75
-)
-
-// Option is a functional option for Engine.
+// Option configures an Engine during construction.
 type Option func(*Engine)
 
 // WithKeyboard sets the keyboard weights.
@@ -173,433 +152,6 @@ func NewEngine(opts ...Option) *Engine {
 
 	return e
 }
-
-// EnableDebug enables per-fix debug tracing.
-func (e *Engine) EnableDebug() {
-	if e != nil {
-		e.debugEnabled = true
-	}
-}
-
-// DisableDebug disables per-fix debug tracing.
-func (e *Engine) DisableDebug() {
-	if e != nil {
-		e.debugEnabled = false
-		e.currentDebug = nil
-	}
-}
-
-// Fix attempts to fix the given command.
-// stderr is optional and used for error parsing.
-func (e *Engine) Fix(cmd, stderr string) itypes.FixResult {
-	return e.FixWithContext(itypes.ParserContext{
-		Command: cmd,
-		Stderr:  stderr,
-	})
-}
-
-// FixWithContext attempts to fix the given command with parser context.
-func (e *Engine) FixWithContext(input itypes.ParserContext) itypes.FixResult {
-	input.Command = strings.TrimSpace(input.Command)
-	debugInfo := e.beginDebugTrace(input)
-	startedAt := time.Now()
-	defer e.clearDebugTrace()
-
-	if input.Command == "" {
-		return e.attachDebug(itypes.FixResult{Fixed: false}, debugInfo, startedAt)
-	}
-
-	if len(input.AliasContext) > 0 {
-		return e.attachDebug(e.fixWithAliasContext(input), debugInfo, startedAt)
-	}
-
-	return e.attachDebug(e.fixWithoutAliasContext(input), debugInfo, startedAt)
-}
-
-// FixCandidatesWithContext returns selectable correction candidates for the input.
-func (e *Engine) FixCandidatesWithContext(input itypes.ParserContext, limit int) []itypes.FixCandidate {
-	input.Command = strings.TrimSpace(input.Command)
-	if e == nil || input.Command == "" || limit < 1 {
-		return nil
-	}
-
-	candidates := make([]itypes.FixCandidate, 0, limit)
-	seen := make(map[string]bool, limit)
-	addResult := func(result itypes.FixResult) {
-		if len(candidates) >= limit || !isMeaningfulFix(input.Command, result) || seen[result.Command] {
-			return
-		}
-		seen[result.Command] = true
-		candidates = append(candidates, itypes.FixCandidate{
-			Command: result.Command,
-			Source:  result.Source,
-			Message: result.Message,
-		})
-	}
-
-	addResult(e.FixWithContext(input))
-	for _, result := range e.distanceFixCandidates(input.Command, limit) {
-		addResult(result)
-		if len(candidates) >= limit {
-			break
-		}
-	}
-
-	return candidates
-}
-
-func (e *Engine) beginDebugTrace(input itypes.ParserContext) *itypes.FixDebugInfo {
-	if e == nil || !e.debugEnabled {
-		return nil
-	}
-
-	debugInfo := &itypes.FixDebugInfo{
-		InputCommand:         input.Command,
-		AliasContextProvided: len(input.AliasContext) > 0,
-		AliasContextEntries:  len(input.AliasContext),
-	}
-	e.currentDebug = debugInfo
-	return debugInfo
-}
-
-func (e *Engine) clearDebugTrace() {
-	if e != nil {
-		e.currentDebug = nil
-	}
-}
-
-func (e *Engine) attachDebug(result itypes.FixResult, debugInfo *itypes.FixDebugInfo, startedAt time.Time) itypes.FixResult {
-	if debugInfo != nil {
-		debugInfo.EngineDuration = time.Since(startedAt)
-		result.Debug = debugInfo
-	}
-	return result
-}
-
-func (e *Engine) debugTrace() *itypes.FixDebugInfo {
-	if e == nil {
-		return nil
-	}
-	return e.currentDebug
-}
-
-func (e *Engine) markDebugFeature(stage string) {
-	debug := e.debugTrace()
-	if debug == nil {
-		return
-	}
-
-	switch stage {
-	case "alias":
-		debug.UsedAlias = true
-		debug.AliasContextUsed = true
-	case fixSourceParser:
-		debug.UsedParser = true
-	case fixSourceHistory:
-		debug.UsedHistory = true
-	case "rule":
-		debug.UsedRule = true
-	case "tree":
-		debug.UsedCommandTree = true
-	case "subcommand":
-		debug.UsedSubcommand = true
-	case fixSourceDistance:
-		debug.UsedDistance = true
-	case "env":
-		debug.UsedEnv = true
-	case "option":
-		debug.UsedOption = true
-	}
-}
-
-func (e *Engine) recordAcceptedFix(pass int, before string, result itypes.FixResult) {
-	debug := e.debugTrace()
-	if debug == nil || !result.Fixed {
-		return
-	}
-
-	e.markDebugFeature(result.Source)
-	if result.UsedParser {
-		debug.UsedParser = true
-	}
-
-	debug.Events = append(debug.Events, itypes.FixDebugEvent{
-		Pass:    pass,
-		Stage:   result.Source,
-		Before:  before,
-		After:   result.Command,
-		Message: result.Message,
-	})
-}
-
-func (e *Engine) recordRejectedCandidate(stage, input, candidate string, distance int, similarity float64, reason string) {
-	debug := e.debugTrace()
-	if debug == nil || input == "" || candidate == "" || reason == "" {
-		return
-	}
-
-	for _, existing := range debug.RejectedCandidates {
-		if existing.Stage == stage && existing.Input == input && existing.Candidate == candidate {
-			return
-		}
-	}
-	if len(debug.RejectedCandidates) >= 5 {
-		return
-	}
-
-	debug.RejectedCandidates = append(debug.RejectedCandidates, itypes.FixDebugCandidate{
-		Stage:      stage,
-		Input:      input,
-		Candidate:  candidate,
-		Distance:   distance,
-		Similarity: similarity,
-		Reason:     reason,
-	})
-}
-
-func (e *Engine) fixWithoutAliasContext(input itypes.ParserContext) itypes.FixResult {
-	input.Command = strings.TrimSpace(input.Command)
-	if input.Command == "" {
-		return itypes.FixResult{Fixed: false}
-	}
-
-	originalCmd := input.Command
-	currentCmd := input.Command
-	messages := make([]string, 0)
-	lastSource := ""
-	resultKind := ""
-	usedParser := false
-
-	passes := e.maxFixPasses
-	if passes < 1 {
-		passes = 1
-	}
-
-	for pass := 1; pass <= passes; pass++ {
-		input.Command = currentCmd
-		result := e.fixOnePass(input)
-		if !isMeaningfulFix(currentCmd, result) {
-			break
-		}
-
-		e.recordAcceptedFix(pass, currentCmd, result)
-		currentCmd = result.Command
-		lastSource = result.Source
-		if result.Message != "" && !slices.Contains(messages, result.Message) {
-			messages = append(messages, result.Message)
-		}
-		if resultKind == "" && result.Kind != "" {
-			resultKind = result.Kind
-		}
-		if result.Source == fixSourceParser {
-			usedParser = true
-			// stderr only belongs to the failed command that triggered this fix.
-			// Once a parser fix lands, later passes must not consume the same stderr again.
-			input.Stderr = ""
-		}
-	}
-
-	if currentCmd != originalCmd {
-		return itypes.FixResult{
-			Fixed:      true,
-			Command:    currentCmd,
-			Source:     lastSource,
-			Message:    strings.Join(messages, "; "),
-			Kind:       resultKind,
-			UsedParser: usedParser,
-		}
-	}
-
-	return itypes.FixResult{Fixed: false}
-}
-
-func (e *Engine) fixOnePass(input itypes.ParserContext) itypes.FixResult {
-	cmd := input.Command
-
-	// 1. Try error parser first (if stderr provided)
-	if input.Stderr != "" {
-		if result := e.tryParser(input); isMeaningfulFix(cmd, result) {
-			return result
-		}
-	}
-
-	// 2. Try explicit user rules
-	if result := e.tryUserRules(cmd); isMeaningfulFix(cmd, result) {
-		return result
-	}
-
-	// 3. Try history
-	if result := e.tryHistory(cmd); isMeaningfulFix(cmd, result) {
-		return result
-	}
-
-	// 4. Try command tree fix
-	if result := e.tryCommandTreeFix(cmd); isMeaningfulFix(cmd, result) {
-		return result
-	}
-
-	// 5. Try builtin rules
-	if result := e.tryBuiltinRules(cmd); isMeaningfulFix(cmd, result) {
-		return result
-	}
-
-	// 6. Try subcommand fix
-	if result := e.trySubcommandFix(cmd); isMeaningfulFix(cmd, result) {
-		return result
-	}
-
-	// 7. Try edit distance
-	if result := e.tryDistance(cmd); isMeaningfulFix(cmd, result) {
-		return result
-	}
-
-	// 8. Try environment variable fix
-	if result := e.tryEnvVarFix(cmd, input.AliasContext); isMeaningfulFix(cmd, result) {
-		return result
-	}
-
-	// 9. Try tool global option fix
-	if e.longOptionFixEnabled {
-		if result := e.tryToolOptionFix(cmd); isMeaningfulFix(cmd, result) {
-			return result
-		}
-	}
-
-	return itypes.FixResult{Fixed: false}
-}
-
-// FixCommand attempts to fix only the command word, preserving arguments.
-func (e *Engine) FixCommand(cmd string) itypes.FixResult {
-	cmd = strings.TrimSpace(cmd)
-	if cmd == "" {
-		return itypes.FixResult{Fixed: false}
-	}
-
-	if result := e.fixCommandWordWithShell(cmd); result.Fixed {
-		return result
-	}
-
-	// Split into command and args
-	parts := strings.Fields(cmd)
-	if len(parts) == 0 {
-		return itypes.FixResult{Fixed: false}
-	}
-
-	cmdWord := parts[0]
-	args := parts[1:]
-
-	// Try to fix just the command word
-	if result := e.tryUserRules(cmdWord); result.Fixed {
-		rebuilt := e.rebuildCommand(result.Command, args, "rule")
-		if isMeaningfulFix(cmd, rebuilt) {
-			return rebuilt
-		}
-	}
-
-	if result := e.tryHistory(cmdWord); result.Fixed {
-		rebuilt := e.rebuildCommand(result.Command, args, fixSourceHistory)
-		if isMeaningfulFix(cmd, rebuilt) {
-			return rebuilt
-		}
-	}
-
-	if replacement := e.findCommandTreeRootForArgs(cmdWord, args); replacement != "" {
-		rebuilt := e.rebuildCommand(replacement, args, "tree")
-		if isMeaningfulFix(cmd, rebuilt) {
-			return rebuilt
-		}
-	}
-
-	if result := e.tryBuiltinRules(cmdWord); result.Fixed {
-		rebuilt := e.rebuildCommand(result.Command, args, "rule")
-		if isMeaningfulFix(cmd, rebuilt) {
-			return rebuilt
-		}
-	}
-
-	if result := e.tryDistance(cmdWord); result.Fixed {
-		rebuilt := e.rebuildCommand(result.Command, args, fixSourceDistance)
-		if isMeaningfulFix(cmd, rebuilt) {
-			return rebuilt
-		}
-	}
-
-	return itypes.FixResult{Fixed: false}
-}
-
-func (e *Engine) fixCommandWordWithShell(cmd string) itypes.FixResult {
-	lines, err := parseShellCommandLines(cmd)
-	if err != nil {
-		return itypes.FixResult{Fixed: false}
-	}
-
-	for _, line := range lines {
-		cmdWord := line.commandWord()
-
-		if rule, ok := e.rules.MatchUser(cmdWord); ok {
-			result := itypes.FixResult{
-				Fixed:   true,
-				Command: line.replaceCommandWord(rule.To),
-				Source:  "rule",
-			}
-			if isMeaningfulFix(cmd, result) {
-				return result
-			}
-		}
-
-		if entry, ok := e.history.Lookup(cmdWord); ok {
-			result := itypes.FixResult{
-				Fixed:   true,
-				Command: line.replaceCommandWord(entry.To),
-				Source:  fixSourceHistory,
-			}
-			if isMeaningfulFix(cmd, result) {
-				return result
-			}
-		}
-
-		args := make([]string, 0, len(line.args)-line.commandIdx-1)
-		for i := line.commandIdx + 1; i < len(line.args); i++ {
-			args = append(args, line.args[i].Lit())
-		}
-
-		if replacement := e.findCommandTreeRootForArgs(cmdWord, args); replacement != "" {
-			result := itypes.FixResult{
-				Fixed:   true,
-				Command: line.replaceCommandWord(replacement),
-				Source:  "tree",
-			}
-			if isMeaningfulFix(cmd, result) {
-				return result
-			}
-		}
-
-		if rule, ok := e.rules.MatchBuiltin(cmdWord); ok {
-			result := itypes.FixResult{
-				Fixed:   true,
-				Command: line.replaceCommandWord(rule.To),
-				Source:  "rule",
-			}
-			if isMeaningfulFix(cmd, result) {
-				return result
-			}
-		}
-
-		if replacement := e.findClosestCommand(cmdWord); replacement != "" {
-			result := itypes.FixResult{
-				Fixed:   true,
-				Command: line.replaceCommandWord(replacement),
-				Source:  fixSourceDistance,
-			}
-			if isMeaningfulFix(cmd, result) {
-				return result
-			}
-		}
-	}
-
-	return itypes.FixResult{Fixed: false}
-}
-
 func (e *Engine) tryParser(input itypes.ParserContext) itypes.FixResult {
 	lines, err := parseShellCommandLines(input.Command)
 	if err == nil {
@@ -1426,234 +978,6 @@ func (e *Engine) resolveShellCommandLine(line *shellCommandLine) (string, string
 	return resolved, resolved, nil
 }
 
-func (e *Engine) findClosestCommand(cmd string) string {
-	matchCfg := e.distanceMatchConfig()
-	bestMatch, bestDistance := e.closestKnownCommand(cmd)
-	if isGoodCommandDistanceMatch(cmd, bestMatch, bestDistance, matchCfg) {
-		return bestMatch
-	}
-	return ""
-}
-
-func (e *Engine) closestKnownCommand(cmd string) (string, int) {
-	matchCfg := e.distanceMatchConfig()
-	bestMatch, bestDistance := e.closestKnownCommandFromCandidates(cmd, e.availableCommandCandidates(cmd, matchCfg.maxEditDistance))
-	if isGoodCommandDistanceMatch(cmd, bestMatch, bestDistance, matchCfg) || e.commandLoader == nil || e.commandsFullyLoad {
-		return bestMatch, bestDistance
-	}
-
-	// Only scan PATH on demand when builtin or seeded commands cannot produce a good candidate.
-	e.loadCommands()
-	return e.closestKnownCommandFromCandidates(cmd, e.availableCommandCandidates(cmd, matchCfg.maxEditDistance))
-}
-
-func (e *Engine) rankedKnownCommandCandidates(cmd string) []commandCandidate {
-	matchCfg := e.distanceMatchConfig()
-	candidates := e.rankedKnownCommandCandidatesFrom(cmd, e.availableCommandCandidates(cmd, matchCfg.maxEditDistance), matchCfg)
-	if len(candidates) > 0 || e.commandLoader == nil || e.commandsFullyLoad {
-		return candidates
-	}
-
-	e.loadCommands()
-	return e.rankedKnownCommandCandidatesFrom(cmd, e.availableCommandCandidates(cmd, matchCfg.maxEditDistance), matchCfg)
-}
-
-func (e *Engine) rankedKnownCommandCandidatesFrom(cmd string, knownCommands []commandRuneCandidate, matchCfg distanceMatchConfig) []commandCandidate {
-	cmdRunes := []rune(cmd)
-	candidates := make([]commandCandidate, 0, len(knownCommands))
-	seen := make(map[string]bool, len(knownCommands))
-
-	for _, known := range knownCommands {
-		if known.name == "" || seen[known.name] {
-			continue
-		}
-		seen[known.name] = true
-
-		d := distanceRunes(cmdRunes, known.runes, e.keyboard)
-		if !isGoodCommandDistanceMatch(cmd, known.name, d, matchCfg) {
-			continue
-		}
-		candidates = append(candidates, commandCandidate{
-			name:       known.name,
-			distance:   d,
-			similarity: SimilarityFromDistance(len(cmd), len(known.name), d),
-			priority:   e.commandPriority(known.name),
-			transposed: utils.IsSingleAdjacentTransposition(cmd, known.name),
-		})
-	}
-
-	sortCommandCandidates(candidates)
-	return candidates
-}
-
-func (e *Engine) closestKnownCommandFromCandidates(cmd string, knownCommands []commandRuneCandidate) (string, int) {
-	cmdRunes := []rune(cmd)
-	candidates := make([]commandCandidate, 0, len(knownCommands))
-	seen := make(map[string]bool, len(knownCommands))
-
-	for _, known := range knownCommands {
-		if known.name == "" || seen[known.name] {
-			continue
-		}
-		seen[known.name] = true
-
-		d := distanceRunes(cmdRunes, known.runes, e.keyboard)
-		candidates = append(candidates, commandCandidate{
-			name:       known.name,
-			distance:   d,
-			similarity: SimilarityFromDistance(len(cmd), len(known.name), d),
-			priority:   e.commandPriority(known.name),
-			transposed: utils.IsSingleAdjacentTransposition(cmd, known.name),
-		})
-	}
-
-	if len(candidates) == 0 {
-		return "", 999
-	}
-
-	sortCommandCandidates(candidates)
-
-	return candidates[0].name, candidates[0].distance
-}
-
-func (e *Engine) availableCommands() []string {
-	e.ensureAvailableCommandsFresh()
-	return e.availableCmds
-}
-
-func (e *Engine) availableCommandCandidates(cmd string, maxEditDistance int) []commandRuneCandidate {
-	e.ensureAvailableCommandsFresh()
-	return e.availableCmdIndex.candidatesFor(cmd, maxEditDistance)
-}
-
-func (e *Engine) isAvailableCommand(cmd string) bool {
-	e.availableCommands() // ensure cache is fresh
-	_, ok := e.availableCmdsSet[cmd]
-	return ok
-}
-
-func (e *Engine) hasKnownCommand(cmd string) bool {
-	if e.isAvailableCommand(cmd) {
-		return true
-	}
-	if e.commandLoader == nil || e.commandsFullyLoad {
-		return false
-	}
-
-	e.loadCommands()
-	return e.isAvailableCommand(cmd)
-}
-
-func (e *Engine) loadCommands() {
-	e.commandsLoadOnce.Do(func() {
-		if e.commandLoader == nil {
-			e.refreshAvailableCommands()
-			e.commandsFullyLoad = true
-			return
-		}
-
-		loaded := e.commandLoader()
-		if debug := e.debugTrace(); debug != nil {
-			debug.LoadedPATHCommands = true
-			debug.LoadedPATHCommandCount = len(loaded)
-		}
-		e.setCommands(utils.MergeUniqueStrings(e.commands, e.filterDisabledCommands(loaded)...))
-		e.refreshAvailableCommands()
-		e.commandsFullyLoad = true
-	})
-}
-
-func (e *Engine) setCommands(commands []string) {
-	e.commands = append([]string(nil), commands...)
-	e.commandsVersion++
-}
-
-func (e *Engine) addCommands(commands ...string) {
-	if len(commands) == 0 {
-		return
-	}
-	e.commands = append(e.commands, commands...)
-	e.commandsVersion++
-}
-
-func (e *Engine) ensureAvailableCommandsFresh() {
-	if e.availableCmdsVersion != e.commandsVersion {
-		e.refreshAvailableCommands()
-	}
-}
-
-func (e *Engine) refreshAvailableCommands() {
-	e.availableCmds = e.filterDisabledCommands(e.commands)
-	e.availableCmdRunes = commandRuneCandidatesFromStrings(e.availableCmds)
-	e.availableCmdIndex = newCommandCandidateIndex(e.availableCmdRunes)
-	e.availableCmdsSet = make(map[string]struct{}, len(e.availableCmds))
-	for _, cmd := range e.availableCmds {
-		e.availableCmdsSet[cmd] = struct{}{}
-	}
-	e.availableCmdsVersion = e.commandsVersion
-}
-
-func commandRuneCandidatesFromStrings(commands []string) []commandRuneCandidate {
-	candidates := make([]commandRuneCandidate, 0, len(commands))
-	for _, cmd := range commands {
-		candidates = append(candidates, commandRuneCandidate{
-			name:  cmd,
-			runes: []rune(cmd),
-		})
-	}
-	return candidates
-}
-
-func newCommandCandidateIndex(candidates []commandRuneCandidate) commandCandidateIndex {
-	index := commandCandidateIndex{
-		all:       candidates,
-		byRuneLen: make(map[int][]commandRuneCandidate),
-	}
-	if len(candidates) == 0 {
-		return index
-	}
-
-	index.minLen = len(candidates[0].runes)
-	index.maxLen = index.minLen
-	for _, candidate := range candidates {
-		runeLen := len(candidate.runes)
-		index.byRuneLen[runeLen] = append(index.byRuneLen[runeLen], candidate)
-		if runeLen < index.minLen {
-			index.minLen = runeLen
-		}
-		if runeLen > index.maxLen {
-			index.maxLen = runeLen
-		}
-	}
-
-	return index
-}
-
-func (index commandCandidateIndex) candidatesFor(cmd string, maxEditDistance int) []commandRuneCandidate {
-	if len(index.all) == 0 {
-		return nil
-	}
-	if maxEditDistance < 0 {
-		return index.all
-	}
-
-	cmdLen := len([]rune(cmd))
-	minLen := cmdLen - maxEditDistance
-	if minLen < 0 {
-		minLen = 0
-	}
-	maxLen := cmdLen + maxEditDistance
-	if minLen <= index.minLen && maxLen >= index.maxLen {
-		return index.all
-	}
-
-	candidates := make([]commandRuneCandidate, 0)
-	for runeLen := minLen; runeLen <= maxLen; runeLen++ {
-		candidates = append(candidates, index.byRuneLen[runeLen]...)
-	}
-	return candidates
-}
-
 func (e *Engine) distanceMatchConfig() distanceMatchConfig {
 	return distanceMatchConfig{
 		keyboard:            e.keyboard,
@@ -1668,20 +992,6 @@ func (e *Engine) longOptionMatchConfig() distanceMatchConfig {
 		cfg.similarityThreshold = longOptionSimilarityThreshold
 	}
 	return cfg
-}
-
-func (e *Engine) filterDisabledCommands(commands []string) []string {
-	if len(e.disabledCommands) == 0 {
-		return commands
-	}
-
-	filtered := make([]string, 0, len(commands))
-	for _, command := range commands {
-		if !e.disabledCommands[command] {
-			filtered = append(filtered, command)
-		}
-	}
-	return filtered
 }
 
 func (e *Engine) isProtectedCommandWord(cmd string) bool {
@@ -2243,86 +1553,6 @@ func (e *Engine) ListHistory() []itypes.HistoryEntry {
 // RecordHistory records a correction that actually happened.
 func (e *Engine) RecordHistory(from, to string) error {
 	return e.history.Record(from, to)
-}
-
-type commandCandidate struct {
-	name       string
-	distance   int
-	similarity float64
-	priority   int
-	transposed bool
-}
-
-func sortCommandCandidates(candidates []commandCandidate) {
-	sort.Slice(candidates, func(i, j int) bool {
-		// Exact matches must stay first; otherwise, prefer adjacent
-		// transpositions over ordinary fuzzy matches from PATH.
-		if cmp := compareFuzzyCandidateOrder(
-			candidates[i].distance, candidates[j].distance,
-			candidates[i].transposed, candidates[j].transposed,
-			candidates[i].similarity, candidates[j].similarity,
-		); cmp != 0 {
-			return cmp < 0
-		}
-		if candidates[i].priority != candidates[j].priority {
-			return candidates[i].priority > candidates[j].priority
-		}
-		return candidates[i].name < candidates[j].name
-	})
-}
-
-func compareFuzzyCandidateOrder(distanceA, distanceB int, transposedA, transposedB bool, similarityA, similarityB float64) int {
-	if distanceA == 0 || distanceB == 0 {
-		switch {
-		case distanceA < distanceB:
-			return -1
-		case distanceA > distanceB:
-			return 1
-		default:
-			return 0
-		}
-	}
-	if transposedA != transposedB {
-		if transposedA {
-			return -1
-		}
-		return 1
-	}
-	if distanceA != distanceB {
-		if distanceA < distanceB {
-			return -1
-		}
-		return 1
-	}
-	if similarityA != similarityB {
-		if similarityA > similarityB {
-			return -1
-		}
-		return 1
-	}
-	return 0
-}
-
-func (e *Engine) commandPriority(cmd string) int {
-	score := e.rules.TargetPriority(cmd)
-
-	if commands.IsCommonCommand(cmd) {
-		score += 50
-	}
-
-	if commands.IsShellBuiltin(cmd) {
-		score += 25
-	}
-
-	if e.toolTrees != nil && e.toolTrees.HasSubcommands(cmd) {
-		score += 25
-	}
-
-	if e.commandTrees != nil && e.commandTrees.HasRoot(cmd) {
-		score += 50
-	}
-
-	return score
 }
 
 func (e *Engine) clearConflictingHistory(from string) error {
